+++ src/Component/OpenApi3/Tests/fixtures/issue-810/expected/Endpoint/MimeTypeGeneratedValidDocBlock.php
@@ -1,0 +1,62 @@
+<?php
+
+namespace Jane\Component\OpenApi3\Tests\Expected\Endpoint;
+
+class MimeTypeGeneratedValidDocBlock extends \Jane\Component\OpenApi3\Tests\Expected\Runtime\Client\BaseEndpoint implements \Jane\Component\OpenApi3\Tests\Expected\Runtime\Client\Endpoint
+{
+    protected $accept;
+    /**
+     * 
+     *
+     * @param string $requestBody 
+     * @param array $accept Accept content header *\/*|application/json
+     */
+    public function __construct(string $requestBody, array $accept = [])
+    {
+        $this->body = $requestBody;
+        $this->accept = $accept;
+    }
+    use \Jane\Component\OpenApi3\Tests\Expected\Runtime\Client\EndpointTrait;
+    public function getMethod(): string
+    {
+        return 'POST';
+    }
+    public function getUri(): string
+    {
+        return '/test-simple';
+    }
+    public function getBody(\Symfony\Component\Serializer\SerializerInterface $serializer, $streamFactory = null): array
+    {
+        if (is_string($this->body)) {
+            return [['Content-Type' => ['application/json']], json_encode($this->body)];
+        }
+        return [[], null];
+    }
+    public function getExtraHeaders(): array
+    {
+        if (empty($this->accept)) {
+            return ['Accept' => ['*/*', 'application/json']];
+        }
+        return $this->accept;
+    }
+    /**
+     * {@inheritdoc}
+     *
+     *
+     * @return null
+     */
+    protected function transformResponseBody(\Psr\Http\Message\ResponseInterface $response, \Symfony\Component\Serializer\SerializerInterface $serializer, ?string $contentType = null)
+    {
+        $status = $response->getStatusCode();
+        $body = (string) $response->getBody();
+        if (404 === $status) {
+        }
+        if (is_null($contentType) === false && (200 === $status && mb_strpos($contentType, 'application/json') !== false)) {
+            return json_decode($body);
+        }
+    }
+    public function getAuthenticationScopes(): array
+    {
+        return [];
+    }
+}